// src/app/model/item-attatchment.model.ts
export interface ItemAttatchment {
    companyid: any;
    recId: any;

    InventoryId: any;

    Type: any;

    Explanation: any;

    fileName: any;

    Attachment: any;

    Thumbnail: any;

    InUse: any;

    InsertedAt: any;

    InsertedBy: any;

    UpdatedAt: any;

    UpdatedBy: any;

    IsDeleted: any;

    DeletedAt: any;

    DeletedBy: any;

    Uuid: any;
}